refactor: バリデーションの格納場所を変更
