+++ presentation/src/main/kotlin/com/bookmanagementsystem/presentation/model/BookResponseModel.kt
@@ -1,12 +1,11 @@
 package com.bookmanagementsystem.presentation.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Min
-import jakarta.validation.constraints.Max
-import jakarta.validation.constraints.Size
 import jakarta.validation.Valid
-import com.bookmanagementsystem.usecase.validation.PastOnly
+import jakarta.validation.constraints.Max
+import jakarta.validation.constraints.Min
+import jakarta.validation.constraints.NotNull
+import jakarta.validation.constraints.Size
 
 /**
  * 書籍のAPIで汎用的に使用するレスポンス情報
