+++ presentation/src/main/kotlin/com/bookmanagementsystem/presentation/model/CreateAuthorRequestModel.kt
@@ -1,12 +1,9 @@
 package com.bookmanagementsystem.presentation.model
 
+import com.bookmanagementsystem.usecase.author.register.PastOnly
 import com.fasterxml.jackson.annotation.JsonProperty
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Min
-import jakarta.validation.constraints.Max
-import jakarta.validation.constraints.Size
 import jakarta.validation.Valid
-import com.bookmanagementsystem.usecase.validation.PastOnly
+import jakarta.validation.constraints.NotNull
 
 /**
  * 著者登録APIのリクエスト情報
@@ -20,7 +17,7 @@ data class CreateAuthorRequestModel(
 
     /* 生年月日 */
     @field:Valid
-    @field:com.bookmanagementsystem.usecase.validation.PastOnly
+    @field:PastOnly
     @get:JsonProperty("birthDate") val birthDate: java.time.LocalDate?
 ) {
