+++ presentation/src/main/kotlin/com/bookmanagementsystem/presentation/model/ErrorModel.kt
@@ -1,12 +1,6 @@
 package com.bookmanagementsystem.presentation.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Min
-import jakarta.validation.constraints.Max
-import jakarta.validation.constraints.Size
-import jakarta.validation.Valid
-import com.bookmanagementsystem.usecase.validation.PastOnly
 
 /**
  * エラー情報
