+++ presentation/src/main/kotlin/com/bookmanagementsystem/presentation/model/NotFoundErrorResponseModel.kt
@@ -1,12 +1,7 @@
 package com.bookmanagementsystem.presentation.model
 
 import com.fasterxml.jackson.annotation.JsonProperty
-import jakarta.validation.constraints.NotNull
-import jakarta.validation.constraints.Min
-import jakarta.validation.constraints.Max
-import jakarta.validation.constraints.Size
 import jakarta.validation.Valid
-import com.bookmanagementsystem.usecase.validation.PastOnly
 
 /**
  * 汎用的なエラーレスポンス
