+++ usecase/src/main/kotlin/com/bookmanagementsystem/usecase/book/register/BookRegisterCommand.kt
@@ -5,7 +5,6 @@ import com.bookmanagementsystem.domain.book.Book
 import com.bookmanagementsystem.domain.book.BookPrice
 import com.bookmanagementsystem.domain.book.BookPublishStatus
 import com.bookmanagementsystem.domain.core.ID
-import com.bookmanagementsystem.usecase.validation.CanChangeBookPublishStatus
 
 sealed interface BookRegisterCommand
